+++ system/src/system/FalloutHovers.mjs
@@ -2,15 +2,23 @@ export class FalloutHovers {
 	static LIST = {};
 
 	static async loadList() {
+		fallout.logger.debug("Loading tooltips");
+
 		const listLocation = await game.settings.get("fallout", "hoversJsonLocation");
 		const jsonFile = await fetch(listLocation);
 		const content = await jsonFile.json();
 
 		for await (const key of Object.keys(content)) {
 			let qEnriched = await TextEditor.enrichHTML(content[key], {async: true});
-			content[key] = qEnriched.replaceAll("&", "&amp;").replaceAll("<", "&lt;").replaceAll(">", "&gt;").replaceAll('"', "&quot;").replaceAll("'", "&#039;");
+
+			content[key] = qEnriched
+				.replaceAll("&", "&amp;")
+				.replaceAll("<", "&lt;")
+				.replaceAll(">", "&gt;")
+				.replaceAll('"', "&quot;")
+				.replaceAll("'", "&#039;");
 		}
-		// console.warn(content)
+
 		this.LIST = content;
 	}
 }