+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [#33] Merged Polish language updates from Crowdin
 - [#35] Remove usage of deprecated `ActiveEffect._getSourceName()`
 - [#36] Remove usage of deprecated `ActiveEffect#label`
+- [#46] Implement schema migration tools ready for likely schema changes
 
 ## v11.2.6